v0.02.02-run: archival snapshot of the validated V100 Stage-1 run
Preserves, verbatim, the as-run code + artifacts that produced the validated
2000-step run on V100 (R1-Distill-Qwen-1.5B frozen backbone, σ held throughout,
γ -> 0.083). This is the PRE-rewrite lineage; the Aeon-only from-scratch rewrite
lives on V0.02.03 and is unaffected.

Overlaid Dylan's actual run files onto the committed V0.02.02 base:
- aeon/transformer.py, aeon/substrate/rwkv_cell.py, scripts/train.py — as-run
  versions (include the per-step gamma.grad debug logging used during the run)
- chat.py — inference helper (loads the trained ckpt; R1 tokenizer/backbone)
- runs/stage1_hybrid/train.log — full 2000-step training log

Note: this branch intentionally contains R1/Qwen/transformers references — it is
the historical record of the validated run, not the Aeon-only line (V0.02.03).
The trained checkpoint (ckpt_2000.pt) is held locally by Dylan, not in the repo.

+++ aeon/transformer.py
@@ -308,7 +308,7 @@ class HybridTransformer(nn.Module):
         # write surface: manifold (H_rec) -> hidden (D), γ-gated from 0 (warm start)
         self.write_proj = nn.Linear(h_rec, self.D, bias=False)
         nn.init.normal_(self.write_proj.weight, std=0.02)
-        self.gamma = nn.Parameter(torch.zeros(1))
+        self.gamma = nn.Parameter(torch.zeros(1, dtype=torch.float32))
 
     # ---- weight init from R1 ---------------------------------------------
     def load_pretrained(self, checkpoint_dir: str) -> dict:
@@ -330,7 +330,7 @@ class HybridTransformer(nn.Module):
         return self.read_proj(hidden)
 
     def inject(self, hidden, signal):
-        return hidden + self.gamma * self.write_proj(signal)
+        return (hidden.float() + self.gamma * self.write_proj(signal).float()).to(hidden.dtype)
 
     def logits(self, hidden):
         return self.model.logits(hidden)
